Classes outside the MainApp file.
