+++ src/main/java/core/basesyntax/Machine.java
@@ -1,0 +1,6 @@
+package core.basesyntax;
+
+abstract class Machine {
+    public abstract void doWork();
+    public abstract void stopWork();
+}
